# packages/api-spec/openapi.yaml
openapi: "3.1.0"

info:
  title: Kittens Game API
  version: "0.1.0"
  description: |
    HTTP + WebSocket API for the Kittens Game modern rewrite.
    The OpenAPI spec is the source of truth for the client/server boundary.
    All request/response shapes are validated by Zod schemas in `src/schemas.ts`.

servers:
  - url: http://localhost:3000
    description: Local development

paths:
  /api/health:
    get:
      operationId: getHealth
      summary: Health check
      description: Returns server status and version. Used by CI and load balancers.
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/game/state:
    get:
      operationId: getGameState
      summary: Get full game state snapshot
      description: Returns the complete serialized game state. Clients use this on initial connect.
      responses:
        "200":
          description: Full game state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStateResponse"

  /api/game/action:
    post:
      operationId: postGameAction
      summary: Dispatch a game action
      description: |
        Applies a typed action to the current game state via the engine reducer.
        Returns the updated state. The action type is a discriminated union —
        new types are added as later epics are completed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameActionRequest"
      responses:
        "200":
          description: Action applied successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "400":
          description: Invalid action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  /api/game/tick:
    post:
      operationId: postGameTick
      summary: Advance one tick (testing only)
      description: |
        Manually advances the game by one tick. Intended for test automation only.
        Should be disabled or guarded in production builds.
      responses:
        "200":
          description: Tick applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStateResponse"

  /api/game/save:
    get:
      operationId: getGameSave
      summary: Export save data
      description: |
        Returns the full serialized save snapshot as JSON.
        The `saveVersion` field drives migration on import.
      responses:
        "200":
          description: Save data exported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveExportResponse"

  /api/game/load:
    post:
      operationId: postGameLoad
      summary: Import save data
      description: |
        Replaces the current game state with the provided save data.
        The server validates the saveVersion and applies migrations if needed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveImportRequest"
      responses:
        "200":
          description: Save loaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStateResponse"
        "400":
          description: Invalid or unmigratable save data

  /api/game/import-legacy:
    post:
      operationId: postGameImportLegacy
      summary: Import a legacy Kittens Game save
      description: |
        Accepts a legacy KG save string (LZString-compressed base64 or UTF-16,
        or raw JSON). Decompresses, migrates to the new schema, and loads it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyImportRequest"
      responses:
        "200":
          description: Legacy save imported and loaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStateResponse"
        "400":
          description: Decompression failed, invalid JSON, or migration error

  /api/game/reset:
    post:
      operationId: postGameReset
      summary: Reset game state
      description: |
        Soft reset: preserves prestige/paragon (default).
        Hard reset: wipes everything including prestige.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameResetRequest"
      responses:
        "200":
          description: Game reset to initial state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStateResponse"

components:
  schemas:
    HealthResponse:
      type: object
      required: [status, version]
      properties:
        status:
          type: string
          enum: [ok]
        version:
          type: string
          example: "0.1.0"

    GameStateResponse:
      type: object
      required: [version, tick]
      properties:
        version:
          type: integer
          minimum: 1
          description: Save format version
        tick:
          type: integer
          minimum: 0
          description: Monotonically increasing tick counter

    GameActionRequest:
      oneOf:
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [TICK]
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [GATHER_CATNIP]
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [BUY_BUILDING]
            name:
              type: string
        - type: object
          required: [type, job]
          properties:
            type:
              type: string
              enum: [ASSIGN_JOB]
            job:
              type: string
        - type: object
          required: [type, job]
          properties:
            type:
              type: string
              enum: [UNASSIGN_JOB]
            job:
              type: string
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [RESEARCH]
            name:
              type: string
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [RESEARCH_POLICY]
            name:
              type: string
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [PURCHASE_UPGRADE]
            name:
              type: string
        - type: object
          required: [type, name, amount]
          properties:
            type:
              type: string
              enum: [CRAFT]
            name:
              type: string
            amount:
              type: number
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [BUY_ZIGGURAT_UPGRADE]
            name:
              type: string
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [BUY_RELIGION_UPGRADE]
            name:
              type: string
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [BUY_TRANSCENDENCE_UPGRADE]
            name:
              type: string
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [PRAISE]
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [ADORE]
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [TRANSCEND]
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [PURCHASE_PERK]
            name:
              type: string
              description: Perk name to purchase
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [SOFT_RESET]
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [START_CHALLENGE]
            name:
              type: string
              description: Challenge name to start
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [COMPLETE_CHALLENGE]
            name:
              type: string
              description: Challenge name to complete
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [LAUNCH_MISSION]
            name:
              type: string
              description: Space mission program name to launch
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [BUY_SPACE_BUILDING]
            name:
              type: string
              description: Space building name to purchase
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [SEND_EMBASSY]
            name:
              type: string
              description: Race name to send embassy to
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [TRADE]
            name:
              type: string
              description: Race name to trade with
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [BUY_CFU]
            name:
              type: string
              description: Chronoforge upgrade name to buy
        - type: object
          required: [type, name]
          properties:
            type:
              type: string
              enum: [BUY_VSU]
            name:
              type: string
              description: Voidspace upgrade name to buy
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [SHATTER_TC]
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [BURN_PARAGON]
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [SACRIFICE_UNICORNS]
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [SACRIFICE_ALICORNS]
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [REFINE_TIME_CRYSTALS]
        - type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [HUNT]
      discriminator:
        propertyName: type

    ActionResult:
      type: object
      required: [ok, state]
      properties:
        ok:
          type: boolean
        error:
          type: string
          description: Error message when ok is false
        state:
          $ref: "#/components/schemas/GameStateResponse"

    SaveExportResponse:
      type: object
      required: [saveVersion, data]
      properties:
        saveVersion:
          type: integer
          minimum: 1
          description: Save format version number
        data:
          type: object
          additionalProperties: true
          description: Full serialized game state

    SaveImportRequest:
      type: object
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: true

    LegacyImportRequest:
      type: object
      required: [data]
      properties:
        data:
          type: string
          minLength: 1
          description: LZString-compressed (base64 or UTF-16) or raw JSON save string

    GameResetRequest:
      type: object
      properties:
        hard:
          type: boolean
          default: false
          description: If true, wipes prestige/paragon. If false or absent, soft reset only.

  # WebSocket messages are documented here for reference.
  # They are not OpenAPI request/response bodies — they travel over WS /ws.
  # Zod schemas for these live in src/schemas.ts.
  #
  # WsStateDelta: { type: "STATE_DELTA", payload: Partial<GameStateResponse>, ts: number }
  # WsConnected:  { type: "CONNECTED",   payload: { sessionId: string },       ts: number }